Fix lint

+++ core/chain_makers.go
@@ -275,7 +275,7 @@ func GenerateChain(config *params.ChainConfig, parent *types.Block, engine conse
 
 	for i := 0; i < n; i++ {
 		stateReader := state.NewDbStateReader(dbCopy)
-		stateWriter := state.NewDbStateWriter(dbCopy, uint64(parent.Number().Uint64()+uint64(i)+1))
+		stateWriter := state.NewDbStateWriter(dbCopy, parent.Number().Uint64()+uint64(i)+1)
 		ibs := state.New(stateReader)
 		block, receipt, err := genblock(i, parent, ibs, stateReader, stateWriter)
 		if err != nil {
